// libs/shared/modules/src/data/models/domain/controlRequirement.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ControlRequirement {
  /**
   * the id of requirement represetned by this control requirements
   */
  requirement_id?: string;
  /**
   * the name of the requirement represetned by this control requirements
   */
  requirement_name?: string;
  /**
   * is this requirement applicable
   */
  requirement_applicability?: boolean;
  /**
   * Whether the control requirement has a slack pending task or not
   */
  requirement_has_pending_slack_task?: boolean;
  /**
   * Explains how does this requirement proves the control
   */
  requirement_help?: string;
  /**
   * id's of the evidence that can fulfil this evidence group
   */
  evidence_that_fulfil?: Array<string>;
  /**
   * whether this control requirement can be automated. true if at least one of the evidence that fulfil this group is automatable.
   */
  evidence_automation?: boolean;

  /**
   * The name of the user who last added/edited this requirement, this is applicable for custom requirements only.
   */
  readonly requirement_edited_by?: string;
  /**
   * The timestamp of which this requirement was last added/edited at, this is applicable for custom requirements only.
   */
  readonly requirement_last_edit_time?: string | number; // this is tmp until me migrate all dates to be strings

  // These properties are not recieved from the backend at the moment, but will be in the future.
  // For now, their values will be populated when we run mapper on all the controls.

  requirement_evidence_ids?: string[];

  /**
   * specifies whether this requirement is a custom requirement or not
   */
  requirement_is_custom?: boolean;

  /**
   * Indicates whether requirement has note
   */
  requirement_note_exists?: boolean;

  /**
   * List of services that automates this requirement
   */
  services_that_automates?: string[];

  /**
   * The requirement related policies
   */
  requirement_related_policies_ids?: string[];

  snapshot_id?: string;
}